% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC.fcrrp.R
\name{AIC.fcrrp}
\alias{AIC.fcrrp}
\title{Akaike's An Information Criterion}
\usage{
\method{AIC}{fcrrp}(object, ..., k = 2)
}
\arguments{
\item{object}{\code{fcrrp} object}

\item{...}{Additional arguments. Not implemented.}

\item{k}{Numeric, the penalty per parameter to be used; the default k = 2 is the classical AIC.}
}
\value{
A numeric value with the corresponding AIC (or BIC, or ..., depending on \code{k}).
}
\description{
Similar functional utility to \code{coef} methods.
}
