% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.fcrr.R
\name{coef.fcrr}
\alias{coef.fcrr}
\title{Extract coefficients from an "fcrr" object.}
\usage{
\method{coef}{fcrr}(object, ...)
}
\arguments{
\item{object}{\code{fcrr} object}

\item{...}{some methods for this generic function require additional arguments. This package does not.}
}
\description{
Similar functional utility to \code{coef} methods.
}
