% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.fcrrp.R
\name{coef.fcrrp}
\alias{coef.fcrrp}
\title{Extract coefficients from an "fcrrp" object.}
\usage{
\method{coef}{fcrrp}(object, ...)
}
\arguments{
\item{object}{\code{fcrrp} object}

\item{...}{Additional arguments. Not implemented.}
}
\value{
Coefficients extracted from the model object \code{object}.
}
\description{
Similar functional utility to \code{coef} methods.
}
