% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastCrr.R
\name{fastCrr}
\alias{fastCrr}
\title{Fast Fine-Gray Model Estimation}
\usage{
fastCrr(ftime, fstatus, X, failcode = 1, cencode = 0, eps = 1e-06,
  max.iter = 1000, getBreslowJumps = TRUE, standardize = TRUE,
  variance = TRUE, var.control = varianceControl(B = 100,
  useMultipleCores = FALSE), returnDataFrame = FALSE)
}
\arguments{
\item{ftime}{A vector of event/censoring times.}

\item{fstatus}{A vector with unique code for each event type and a separate code for censored observations.}

\item{X}{A matrix of fixed covariates (nobs x ncovs)}

\item{failcode}{Integer: code of \code{fstatus} that event type of interest (default is 1)}

\item{cencode}{Integer: code of \code{fstatus} that denotes censored observations (default is 0)}

\item{eps}{Numeric: algorithm stops when the relative change in any coefficient is less than \code{eps} (default is \code{1E-6})}

\item{max.iter}{Numeric: maximum iterations to achieve convergence (default is 1000)}

\item{getBreslowJumps}{Logical: Output jumps in Breslow estimator for the cumulative hazard.}

\item{standardize}{Logical: Standardize design matrix.}

\item{variance}{Logical: Get standard error estimates for parameter estimates via bootstrap.}

\item{var.control}{List of options for variance estimation.}

\item{returnDataFrame}{Logical: Return (ordered) data frame.}
}
\value{
Returns a list of class \code{fcrr}.
}
\description{
Estimates parameters for the proportional subdistribution hazards model using two-way linear scan approach.
}
\details{
Fits the 'proportional subdistribution hazards' regression model described in Fine and Gray (1999) using a novel two-way linear scan approach.
}
\examples{
library(fastcmprsk)

set.seed(10)
ftime <- rexp(200)
fstatus <- sample(0:2, 200, replace = TRUE)
cov <- matrix(runif(1000), nrow = 200)
dimnames(cov)[[2]] <- c('x1','x2','x3','x4','x5')
fit <- fastCrr(ftime, fstatus, cov, variance = FALSE)

#To parallellize variance estimation (make sure doParallel is loaded)
myClust <- makeCluster(2)
registerDoParallel(myClust)
fit1 <- fastCrr(ftime, fstatus, cov, variance = FALSE,
var.control = varianceControl(B = 100, useMultipleCores = TRUE))
stopCluster(myClust)



}
\references{
Fine J. and Gray R. (1999) A proportional hazards model for the subdistribution of a competing risk.  \emph{JASA} 94:496-509.
}
