% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCIF.R
\name{getCIF}
\alias{getCIF}
\title{Cumulative Incidence Function Estimation}
\usage{
getCIF(fit, cov, getBootstrapVariance = TRUE, B = 100, type = "none",
  alpha = 0.05, seednum = 1991, tL = NULL, tU = NULL, ...)
}
\arguments{
\item{fit}{Output from \code{fcrr} object.}

\item{cov}{A set of covariate values.}

\item{getBootstrapVariance}{Logical: Calculate variance for CIF via bootstrap.}

\item{B}{Number of bootstrap samples for variance estimation.}

\item{type}{Confidence intervals or confidence bands.}

\item{alpha}{Significance level to compute intervals or bands}

\item{seednum}{Seed number of bootstrap sampling.}

\item{tL}{Lower time for band estimation.}

\item{tU}{Upper time for band estimation.}

\item{...}{additional arguments affecting the fastCrr procedure.}
}
\description{
Predicts cumulative incidence function from a \code{fcrr} object.
}
\details{
Calculates the CIF using \code{fcrr} output conditional on \code{cov}.
}
\examples{
set.seed(10)
ftime <- rexp(200)
fstatus <- sample(0:2, 200, replace = TRUE)
cov <- matrix(runif(1000), nrow = 200)
dimnames(cov)[[2]] <- c('x1','x2','x3','x4','x5')
fit <- fastCrr(ftime, fstatus, cov)
cov2 <- rnorm(5)
getCIF(fit, cov2)
}
\references{
Fine J. and Gray R. (1999) A proportional hazards model for the subdistribution of a competing risk.  \emph{JASA} 94:496-509.
}
