% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.fcrr.R
\name{logLik.fcrr}
\alias{logLik.fcrr}
\title{Extract log-pseudo likelihood from an "fcrr" object.}
\usage{
\method{logLik}{fcrr}(object, ...)
}
\arguments{
\item{object}{\code{fcrr} object}

\item{...}{some methods for this generic function require additional arguments. This package does not.}
}
\description{
Similar functional utility to \code{coef} methods.
}
