% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.fcrrp.R
\name{logLik.fcrrp}
\alias{logLik.fcrrp}
\title{Extract log-pseudo likelihood from an "fcrrp" object.}
\usage{
\method{logLik}{fcrrp}(object, ...)
}
\arguments{
\item{object}{\code{fcrrp} object}

\item{...}{Additional arguments. Not implemented.}
}
\value{
Returns the log-pseudo likelihood of object \code{object}.
}
\description{
Similar functional utility to \code{coef} methods.
}
