% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cif.fcrr.R
\name{plot.cif.fcrr}
\alias{plot.cif.fcrr}
\title{Plots predicted cumulative incidence function}
\usage{
\method{plot}{cif.fcrr}(x, ...)
}
\arguments{
\item{x}{\code{cif.fcrr} object (output from \code{getCIF()})}

\item{...}{additional arguments to \code{print()}}
}
\description{
Plots predicted cumulative incidence function
}
