% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cif.fcrr.R
\name{plot.predict.fcrr}
\alias{plot.predict.fcrr}
\title{Plots predicted cumulative incidence function}
\usage{
\method{plot}{predict.fcrr}(x, ...)
}
\arguments{
\item{x}{\code{predict.fcrr} object (output from \code{predict(fcrr object)})}

\item{...}{additional arguments to \code{plot()}}
}
\description{
Plots predicted cumulative incidence function
}
