% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.fcrr.R
\name{print.summary.fcrr}
\alias{print.summary.fcrr}
\title{Prints summary of a fcrr object}
\usage{
\method{print}{summary.fcrr}(x, digits = max(options()$digits - 4, 3),
  ...)
}
\arguments{
\item{x}{\code{fcrr} object (output from \code{fastCrr()})}

\item{...}{additional arguments to \code{print()}}
}
\description{
Prints summary statistics of a fcrr object
}
\details{
Prints the convergence status,
log-pseudo likelihood, the estimated coefficients, the estimated standard errors, and the two-sided p-values for the test of the individual coefficients equal to 0.
}
