% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printCoefmat.fcrr.R
\name{printCoefmat.fcrr}
\alias{printCoefmat.fcrr}
\title{Print Coefficient Matrices}
\usage{
printCoefmat.fcrr(x, conf.int = TRUE, alpha = 0.05,
  digits = max(options()$digits - 5, 2), ...)
}
\arguments{
\item{x}{\code{fcrr} object (output from \code{fastCrr()})}

\item{conf.int}{Logical. Whether or not to outut confidence intervals.}

\item{alpha}{Significance level of the confidence intervals.}

\item{digits}{Numer of significant difits to round to.}

\item{...}{Additional arguments. Not implemented.}
}
\description{
Generate and print summaries of \code{fastCrr} output.
}
\details{
The summary method produces an ANOVA table for the coefficient estimates of the Fine-Gray model.
}
