% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pshBAR.R
\name{pshBAR}
\alias{pshBAR}
\title{Efficient L0-BAR for the Fine-Gray Model}
\usage{
pshBAR(formula, data, eps = 1e-06, max.iter = 1000,
  getBreslowJumps = TRUE, standardize = TRUE, method = c("iterative",
  "cyclic"), lambda = NULL, alpha = 0, lambda.min.ratio = 0.001,
  nlambda = 25, penalty.factor = rep(1, ncol(X)),
  gamma = switch(penalty, scad = 3.7, 2.7))
}
\arguments{
\item{formula}{a formula object, with the response on the left of a ~ operator, and the terms on the right. The response must be a Crisk object as returned by the \code{Crisk} function.}

\item{data}{a data.frame in which to interpret the variables named in the formula.}

\item{eps}{Numeric: algorithm stops when the relative change in any coefficient is less than \code{eps} (default is \code{1E-6})}

\item{max.iter}{Numeric: maximum iterations to achieve convergence (default is 1000)}

\item{getBreslowJumps}{Logical: Output jumps in Breslow estimator for the cumulative hazard.}

\item{standardize}{Logical: Standardize design matrix.}

\item{lambda}{A user-specified sequence of \code{lambda} values for tuning parameters.}

\item{alpha}{L1/L2 weight for elastic net regression.}

\item{lambda.min.ratio}{Smallest value for \code{lambda}, as a fraction of \code{lambda.max} (if \code{lambda} is NULL).}

\item{nlambda}{Number of \code{lambda} values (default is 25).}

\item{penalty.factor}{A vector of weights applied to the penalty for each coefficient. Vector must be of length equal to the number of columns in \code{X}.}

\item{gamma}{Tuning parameter for the MCP/SCAD penalty. Default is 2.7 for MCP and 3.7 for SCAD and should be left unchanged.}

\item{penalty}{Character: Penalty to be applied to the model. Options are "lasso", "scad", "ridge", "mcp", and "enet".}
}
\value{
Returns a list of class \code{fcrrp}.
}
\description{
Fits broken adaptive ridge regression for Fine-Gray regression.
}
\details{
The \code{pshBAR} function penalizes the log-partial likelihood of the proportional subdistribution hazards model
from Fine and Gray (1999) with the Broken Adaptive Ridge (BAR) penalty. A cyclic coordinate descent algorithm is used for implementation.
For stability, the covariate matrix \code{X} is standardized prior to implementation.

Special cases: Fixing \code{xi} and \code{lambda} to 0 results in the standard competing risk regression using \code{crr}.
Fixing \code{lambda} to 0 and specifying \code{xi} will result in a ridge regression solution.
}
\examples{
library(fastcmprsk)
set.seed(10)
ftime <- rexp(200)
fstatus <- sample(0:2, 200, replace = TRUE)
cov <- matrix(runif(1000), nrow = 200)
dimnames(cov)[[2]] <- c('x1','x2','x3','x4','x5')
fit <- fastCrrp(Crisk(ftime, fstatus) ~ cov, lambda = 1, penalty = "RIDGE")
fit$coef
}
\references{
Fu, Z., Parikh, C.R., Zhou, B. (2017) Penalized variable selection in competing risks
regression. \emph{Lifetime Data Analysis} 23:353-376.

Breheny, P. and Huang, J. (2011) Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection. \emph{Ann. Appl. Statist.}, 5: 232-253.

Fine J. and Gray R. (1999) A proportional hazards model for the subdistribution of a competing risk.  \emph{JASA} 94:496-509.
}
