% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.fcrr.R
\name{residuals.fcrr}
\alias{residuals.fcrr}
\title{Calculate residuals for 'fcrr' fit}
\usage{
\method{residuals}{fcrr}(object, type = "martingale", ...)
}
\arguments{
\item{object}{Output from \code{fcrr} object.}

\item{type}{Character string indicating the type of residual desired. Possible values are \code{"cox-snell"} or \code{"martingale"}
which are based under a similar derivation to the Cox proportional hazards model.}

\item{...}{additional arguments (currently unused).}
}
\description{
Calculates martingale and Cox-Snell residuaks for the Fine-Gray model.
}
\details{
Calculates residuals. \code{returnDataFrame = TRUE} must be specified in \code{fastCrr}.
}
\examples{
set.seed(10)
ftime <- rexp(200)
fstatus <- sample(0:2, 200, replace = TRUE)
cov <- matrix(runif(1000), nrow = 200)
dimnames(cov)[[2]] <- c('x1','x2','x3','x4','x5')
fit <- fastCrr(Crisk(ftime, fstatus) ~ cov, returnDataFrame = TRUE)
residuals(fit, type = "martingale")
}
\references{
Fine J. and Gray R. (1999) A proportional hazards model for the subdistribution of a competing risk.  \emph{JASA} 94:496-509.
}
