% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateSparseFineGrayModel.R
\name{simulateSparseFineGrayModel}
\alias{simulateSparseFineGrayModel}
\title{Simulate data from the Fine-Gray Model using a Sparse Design Matrix}
\usage{
simulateSparseFineGrayModel(nobs = 1000, beta1, beta2, p = 0.5,
  u.min = 0, u.max = 1, numCovsPerObs = 0.1, parallel = TRUE,
  ncores = NULL, seednum = 1234)
}
\arguments{
\item{nobs}{Integer: Number of observations in simulated dataset.}

\item{beta1}{A vector of effect sizes for cause 1 of length ncovs}

\item{beta2}{A vector of effect sizes for cause 2 of length ncovs}

\item{p}{Numeric: value between 0 and 1 which controls the mixture probability.}

\item{u.min}{Numeric: controls lower bound of censoring distribution where C ~ U(u.min, u.max)}

\item{u.max}{Numeric: controls upper bound of censoring distribution where C ~ U(u.min, u.max)}

\item{numCovsPerObs}{Numeric: value between 0 and 1 which controls the proportion of non-zero covariate values per observation.}

\item{parallel:}{Logical: Specifies whether to parallelize the construction of the design matrix.}

\item{ncores:}{Integer: Number of cores to use when \code{parallel} is TRUE.}

\item{seednum:}{Integer: For reproducibility.}
}
\value{
Returns a list with \code{outcomes}, \code{covariates}.
}
\description{
Simulate data from the model proposed in Fine and Gray (1999) for two causes. Cause 1 is assumed
to be of primary importance. Design matrix is stored in coordinate format.
}
\details{
The function simulates data according to the setup by Fine and Gray (1999). See their paper for more information.
The \code{outcomes} object stores the row (observation) ID, observed failure time, and censoring indicator, respectively.
The \code{covariates} object stores the row ID, column ID, and value for the non-zero entry in the design matrix, respectively. (e.g. If x_14 = 3, then rowId = 1, covariateId = 4, and covariateValue = 3)
}
\examples{
set.seed(10)
nobs <- 500
beta1 <- c(0.5, 0.4, 0, 0, 0.35, 0, 0, 0.7)
beta2 <- -beta1
X <- matrix(rnorm(nobs * 8), nrow = nobs)
dat <- simulateTwoCauseFineGrayModel(nobs, beta1, beta2, X, u.min = 0, u.max = 1, p = 0.5)
}
\references{
Fine J. and Gray R. (1999) A proportional hazards model for the subdistribution of a competing risk.  \emph{JASA} 94:496-509.
}
