% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.fcrr.R
\name{summary.fcrr}
\alias{summary.fcrr}
\title{Prints summary of a fcrr object}
\usage{
\method{summary}{fcrr}(x, ...)
}
\arguments{
\item{x}{\code{fcrr} object (output from \code{fastCrr()})}

\item{...}{additional arguments to \code{print()}}
}
\description{
Prints summary statistics of a fcrr object
}
\details{
Prints the convergence status,
log-pseudo likelihood, the estimated coefficients, the estimated standard errors, and the two-sided p-values for the test of the individual coefficients equal to 0.
}
