% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.fcrr.R
\name{coef.fcrr}
\alias{coef.fcrr}
\title{Extract coefficients from an "fcrr" object.}
\usage{
\method{coef}{fcrr}(object, ...)
}
\arguments{
\item{object}{\code{fcrr} object}

\item{...}{Additional arguments. Not implemented.}
}
\value{
Coefficients extracted from the model object \code{object}.
}
\description{
Similar functional utility to \code{coef} methods.
}
