% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.fcrr.R
\name{confint.fcrr}
\alias{confint.fcrr}
\title{Confidence Intervals for Model Parameters}
\usage{
\method{confint}{fcrr}(object, parm, level = 0.95, digits = max(options()$digits - 5, 2), ...)
}
\arguments{
\item{object}{\code{fcrr} object (output from \code{fastCrr()})}

\item{parm}{a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.}

\item{level}{the confidence level required}

\item{digits}{Number of significant difits to round to.}

\item{...}{Additional arguments. Not implemented.}
}
\value{
A matrix (or vector) with columns giving lower and upper confidence limits for each coefficient estimate.
}
\description{
Computes confidence intervals for one or more parameters in a fitted model of class \code{fcrr}.
}
\details{
Prints out table of confidence intervals for the Fine-Gray model.
}
