% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.fcrr.R
\name{logLik.fcrr}
\alias{logLik.fcrr}
\title{Extract log-pseudo likelihood from an "fcrr" object.}
\usage{
\method{logLik}{fcrr}(object, ...)
}
\arguments{
\item{object}{\code{fcrr} object}

\item{...}{Additional arguments. Not implemented.}
}
\value{
Returns the log-pseudo likelihood of object \code{object}.
}
\description{
Similar functional utility to \code{coef} methods.
}
