% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fcrrp.R
\name{plot.fcrrp}
\alias{plot.fcrrp}
\title{Plots solution path for penalized methods}
\usage{
\method{plot}{fcrrp}(x, ...)
}
\arguments{
\item{x}{\code{fcrrp} object (output from \code{fastCrrp()})}

\item{...}{additional arguments to \code{plot()}}
}
\description{
Plots solution path for penalized methods
}
