% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.fcrr.R
\name{print.summary.fcrr}
\alias{print.summary.fcrr}
\title{Prints summary of a fcrr x}
\usage{
\method{print}{summary.fcrr}(x, digits = max(options()$digits - 4, 3), ...)
}
\arguments{
\item{x}{output from \code{fastCrr()}.}

\item{digits}{digits for rounding.}

\item{...}{additional arguments to \code{print()}.}
}
\value{
Prints the convergence status,
log-pseudo likelihood, the estimated coefficients, the estimated standard errors, and the two-sided p-values for the test of the individual coefficients equal to 0.
}
\description{
Prints summary statistics of a fcrr x
}
\details{
Prints the convergence status,
log-pseudo likelihood, the estimated coefficients, the estimated standard errors, and the two-sided p-values for the test of the individual coefficients equal to 0.
}
