% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.fcrr.R
\name{summary.fcrr}
\alias{summary.fcrr}
\title{SUmmary method for fastCrr}
\usage{
\method{summary}{fcrr}(x, conf.int = TRUE, alpha = 0.05,
  digits = max(options()$digits - 5, 2), ...)
}
\arguments{
\item{x}{\code{fcrr} x (output from \code{fastCrr()})}

\item{conf.int}{Logical. Whether or not to outut confidence intervals.}

\item{alpha}{Significance level of the confidence intervals.}

\item{digits}{Numer of significant difits to round to.}

\item{...}{additional arguments to \code{print()}}
}
\description{
Generate and print summaries of \code{fastCrr} output.
}
\details{
The summary method produces an ANOVA table for the coefficient estimates of the Fine-Gray model.
}
