% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.fcrr.R
\name{vcov.fcrr}
\alias{vcov.fcrr}
\title{Extract variance-covariance matrix from an "fcrr" object.}
\usage{
\method{vcov}{fcrr}(object, ...)
}
\arguments{
\item{object}{\code{fcrr} object.}

\item{...}{Additional arguments. Not implemented.}
}
\value{
Returns the estimated variance-covariance matrix (via bootstrap) from object \code{object}.
}
\description{
Similar functional utility to \code{vcov} methods.
}
